search bar works, now displays pokemon data and sprites display on page, gives option to add to pokedex, if selected, changes pokemon's pokedex property to true in db
