+++ controllers/api/index.js
@@ -1,5 +1,9 @@
 const router = require('express').Router();
 // Add in routes
 const loginRoutes = require('./loginRoutes.js')
+const searchRoutes = require('./searchRoutes.js')
+
 router.use('/users', loginRoutes)
+router.use('/search', searchRoutes)
+
 module.exports = router;
